+++ src/media/font/art/mod.rs
@@ -1,0 +1,12 @@
+// devela::media::font::art
+//
+//! Ascii-art fonts.
+//
+
+// #[cfg(test)]
+// mod tests;
+
+mod art; // FontArt
+mod block_4_3; // FONT_ART_BLOCK_4_3
+
+pub use {art::*, block_4_3::*};