+++ src/media/font/mod.rs
@@ -5,7 +5,8 @@
 // safety
 #![cfg_attr(feature = "safe_font", forbid(unsafe_code))]
 
-mod bitmap;
+mod art; // FontArt
+mod bitmap; // FontBitmap
 
 // WIPZONE
 // #[cfg(any(feature = "std", feature = "dep_hashbrown"))]
@@ -13,7 +14,10 @@ mod bitmap;
 
 crate::structural_mods! { // _mods, _all
     _mods {
-        pub use super::bitmap::*;
+        pub use super::{
+            art::*,
+            bitmap::*,
+        };
         // WIPZONE
         // #[cfg(feature = "std")]
         // pub use super::bdf::*;